Create BeautifulSoup
Python program that uses the BeautifulSoup library to scrape information from a website. In this example, let's say we're extracting the titles of articles from a blog page. Make sure you have BeautifulSoup installed (pip install beautifulsoup4) before running the program.

Replace "https://example.com/blog" with the actual URL of the website you want to scrape. This program will extract all the titles of articles from the specified webpage and print them out. You can modify it to extract different types of information or from different elements of the webpage as needed.